alacritty: abstract and organize terminal management
