+++ crates/rend3-alacritty/Cargo.toml
@@ -9,6 +9,7 @@ alacritty_terminal = "0.16"
 bytemuck = { version = "1.13", features = ["derive"] }
 font-mud = { path = "../font-mud" }
 glam = "0.20"
+mio-extras = "2"
 owned_ttf_parser = "0.18.1"
 rend3 = "0.3"
 rend3-routine = "0.3"
@@ -17,5 +18,4 @@ wgpu = "^0.12"
 [dev-dependencies]
 image = "0.24"
 rend3-framework = "0.3"
-mio-extras = "2"
 winit = "0.26"
