+++ crates/rend3-alacritty/src/text.rs
@@ -1,0 +1,199 @@
+// Copyright (c) 2023 the Hearth contributors.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::{
+    collections::HashSet,
+    sync::{Arc, Mutex},
+};
+
+use alacritty_terminal::term::cell::Flags;
+use font_mud::glyph_atlas::GlyphAtlas;
+use owned_ttf_parser::{AsFaceRef, OwnedFace};
+use wgpu::{util::DeviceExt, *};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum FontStyle {
+    Regular,
+    Italic,
+    Bold,
+    BoldItalic,
+}
+
+impl FontStyle {
+    pub fn from_cell_flags(flags: Flags) -> Self {
+        if flags.contains(Flags::BOLD_ITALIC) {
+            Self::BoldItalic
+        } else if flags.contains(Flags::ITALIC) {
+            Self::Italic
+        } else if flags.contains(Flags::BOLD) {
+            Self::Bold
+        } else {
+            Self::Regular
+        }
+    }
+}
+
+/// Generic container for all font faces used in a terminal. Eases
+/// the writing of code manipulating all faces at once.
+#[derive(Clone, Debug, Default)]
+pub struct FontSet<T> {
+    pub regular: T,
+    pub italic: T,
+    pub bold: T,
+    pub bold_italic: T,
+}
+
+impl<T> FontSet<T> {
+    pub fn map<O>(self, f: impl Fn(T) -> O) -> FontSet<O> {
+        FontSet {
+            regular: f(self.regular),
+            italic: f(self.italic),
+            bold: f(self.bold),
+            bold_italic: f(self.bold_italic),
+        }
+    }
+
+    pub fn for_each(self, mut f: impl FnMut(T)) {
+        f(self.regular);
+        f(self.italic);
+        f(self.bold);
+        f(self.bold_italic);
+    }
+
+    pub fn get(&self, style: FontStyle) -> &T {
+        match style {
+            FontStyle::Regular => &self.regular,
+            FontStyle::Italic => &self.italic,
+            FontStyle::Bold => &self.bold,
+            FontStyle::BoldItalic => &self.bold_italic,
+        }
+    }
+
+    pub fn get_mut(&mut self, style: FontStyle) -> &mut T {
+        match style {
+            FontStyle::Regular => &mut self.regular,
+            FontStyle::Italic => &mut self.italic,
+            FontStyle::Bold => &mut self.bold,
+            FontStyle::BoldItalic => &mut self.bold_italic,
+        }
+    }
+
+    pub fn zip<O>(self, other: FontSet<O>) -> FontSet<(T, O)> {
+        FontSet {
+            regular: (self.regular, other.regular),
+            italic: (self.italic, other.italic),
+            bold: (self.bold, other.bold),
+            bold_italic: (self.bold_italic, other.bold_italic),
+        }
+    }
+
+    pub fn as_ref(&self) -> FontSet<&T> {
+        FontSet {
+            regular: &self.regular,
+            italic: &self.italic,
+            bold: &self.bold,
+            bold_italic: &self.bold_italic,
+        }
+    }
+
+    pub fn as_mut(&mut self) -> FontSet<&mut T> {
+        FontSet {
+            regular: &mut self.regular,
+            italic: &mut self.italic,
+            bold: &mut self.bold,
+            bold_italic: &mut self.bold_italic,
+        }
+    }
+}
+
+/// A font face and its MSDF glyph atlas.
+pub struct FaceAtlas {
+    pub face: OwnedFace,
+    pub atlas: GlyphAtlas,
+    pub texture: Texture,
+    pub queue: Arc<Queue>,
+    pub touched: Mutex<HashSet<u16>>,
+}
+
+impl FaceAtlas {
+    /// Create a new atlas from a face. Note that this takes time to complete.
+    pub fn new(face: OwnedFace, device: &Device, queue: Arc<Queue>) -> Self {
+        let (atlas, _errors) = GlyphAtlas::new(face.as_face_ref()).unwrap();
+
+        let size = Extent3d {
+            width: atlas.width,
+            height: atlas.height,
+            depth_or_array_layers: 1,
+        };
+
+        let texture = device.create_texture_with_data(
+            &queue,
+            &TextureDescriptor {
+                label: Some("AlacrittyRoutine::glyph_texture"),
+                size,
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: TextureDimension::D2,
+                format: TextureFormat::Rgba8Unorm,
+                usage: TextureUsages::TEXTURE_BINDING | TextureUsages::COPY_DST,
+            },
+            &vec![0u8; (atlas.width * atlas.height * 4) as usize],
+        );
+
+        Self {
+            face,
+            atlas,
+            texture,
+            queue,
+            touched: Default::default(),
+        }
+    }
+
+    /// Generate and upload a glyph bitmap for each glyph that hasn't already been.
+    pub fn touch(&self, glyphs: &[u16]) {
+        let mut touched = self.touched.lock().unwrap();
+        for glyph in glyphs {
+            if touched.insert(*glyph) {
+                let glyph = self.atlas.glyphs.get(*glyph as usize);
+                let Some(Some(glyph)) = glyph else { continue };
+                let bitmap = glyph.shape.generate();
+
+                self.queue.write_texture(
+                    ImageCopyTexture {
+                        texture: &self.texture,
+                        mip_level: 0,
+                        origin: Origin3d {
+                            x: glyph.position.x,
+                            y: glyph.position.y,
+                            z: 0,
+                        },
+                        aspect: TextureAspect::All,
+                    },
+                    bitmap.data_bytes(),
+                    ImageDataLayout {
+                        offset: 0,
+                        bytes_per_row: std::num::NonZeroU32::new(glyph.size.x * 4),
+                        rows_per_image: std::num::NonZeroU32::new(glyph.size.y),
+                    },
+                    Extent3d {
+                        width: glyph.size.x,
+                        height: glyph.size.y,
+                        depth_or_array_layers: 1,
+                    },
+                );
+            }
+        }
+    }
+}
